Crate ic_registry_client

Modules

client
Implementation of the registry client. Calls to the API always return immediately. The provided data provider is polled periodically in the background when start_polling() is called.

fake
A fake implementation of RegistryClient intended to be used in component tests and utility functions where a real registry that polls in the background is not required.

helper
These are Helper traits that wrap a RegistryClient and provide convenience-methods to turn raw bytes into internal data structures. Traits specific to a particular component (crypto comes to mind) will move to the respective crate/component at some point in the future.

nns_registry_replicator
NNS Registry Replicator