Struct ic_registry_client::client::RegistryClientImpl

pub struct RegistryClientImpl { /* fields omitted */ }

Implementations

impl RegistryClientImpl

RegistryClientImpl polls the registry on the NNS subnet and caches the responses.

pub fn new(
data_provider: Arc<dyn RegistryDataProvider>,
metrics_registry: Option<&MetricsRegistry>
) -> Self

Creates a new instance of the RegistryClient.

pub fn fetch_and_start_polling(&self) -> Result<(), RegistryClientError>

Calls poll_once() synchronously, if it succeeds a background task is spawned that continuously polls for updates. The background task is stopped when the object is dropped.

pub fn poll_once(&self) -> Result<(), RegistryClientError>

Fetches the newest updates from the registry data provider.

Returns an error if a poll is already in progress or querying the data provider failed. Returns Ok if querying the data provider succeeded, regardless of whether a newer registry version was available or not.