Trait ic_registry_client::client::RegistryClient

pub trait RegistryClient: Send + Sync {
    fn get_versioned_value(
        &self, 
        key: &str, 
        version: AmountOf<RegistryVersionTag, u64>
    ) -> Result<RegistryVersionedRecord<Vec<u8>>, RegistryClientError>;

    fn get_key_family(
        &self, 
        key_prefix: &str, 
        version: AmountOf<RegistryVersionTag, u64>
    ) -> Result<Vec<String>, RegistryClientError>;

    fn get_latest_version(&self) -> AmountOf<RegistryVersionTag, u64>;

    fn get_value(
        &self, 
        key: &str, 
        version: AmountOf<RegistryVersionTag, u64>
    ) -> Result<Option<Vec<u8>>, RegistryClientError> { ... }
}

The RegistryClient provides methods to query the local state of the registry. All methods on this trait return immediately (i.e. there are no side-effects on the critical path).

Required methods

fn get_versioned_value(
&self,
key: &str,
version: AmountOf<RegistryVersionTag, u64>
) -> Result<RegistryVersionedRecord<Vec<u8>>, RegistryClientError>

The following holds:

(1) ∀ k: get_value(k, get_latest_version()).is_ok()

(A reported version is fully available)

(2) ∀ k: v = get_value(k, ZERO_REGISTRY_VERSION) => (v.is_ok() && v.unwrap().is_none())

(The registry at version zero has a known state.)

(3) ∀ k, t: (v0 = get_value(k, t); v1 = get_value(k, t)) && v0.is_ok() && v1.is_ok() => v0 == v1

(Any two method invocations with the same arguments will always result in equal return values if both calls do not return an error.)

NOTE: The current implementation employs full replication and the cache only grows. Thus, it holds that

∀ k, t <= get_latest_version(): get_value(k, t).is_ok().

However, this might change in the future.

The return value of the function is a serialized protobuf message belonging to the key. The type is opaque and the API does not provide any runtime type information. We might switch to a type such as google.protobuf.Any at some point in the future to provide general runtime type information.

fn get_key_family(
&self,
key_prefix: &str,
version: AmountOf<RegistryVersionTag, u64>
) -> Result<Vec<String>, RegistryClientError>

Returns all keys that start with key_prefix and are present at version version.

Given the definition of get_value above, let K* be the set of all possible keys that start with key_prefix, then the following holds:

(1) ∀ k ∈ K*: get_value(k, version).is_err() <=> get_versioned_key_family(key_prefix, version).is_err()

(For a given version, all keys of a key family must be well-defined.)

(2) ∀ k ∈ K*: get_value(k, version).is_ok().is_some() <=> get_key_family(key_prefix, version).is_ok().contains(k)

(get_key_family is consistent with get_value)

The returned list does not contain any duplicates. There are no guarantees wrt. the order of the contained elements.

fn get_latest_version(&self) -> AmountOf<RegistryVersionTag, u64>

Returns the latest version known to this replica. If the current version of the registry is t, then this method should eventually return a value no less than t.

Loading content...
Provided methods

fn get_value(
&self,
key: &str,
version: AmountOf<RegistryVersionTag, u64>
) -> Result<Option<Vec<u8>>, RegistryClientError>

Loading content...