Trait ic_registry_client::helper::crypto::CryptoRegistry

pub trait CryptoRegistry {
    fn get_crypto_key_for_node(
        &self, 
        node_id: NodeId, 
        key_purpose: KeyPurpose, 
        version: RegistryVersion
    ) -> RegistryClientResult<PublicKeyProto>;

    fn get_threshold_signing_public_key_for_subnet(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<ThresholdSigPublicKey>;

    fn get_tls_certificate(
        &self, 
        node_id: NodeId, 
        version: RegistryVersion
    ) -> RegistryClientResult<X509PublicKeyCert>;

    fn get_initial_dkg_transcripts(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientVersionedResult<DkgTranscripts>;
}

Required methods

fn get_crypto_key_for_node(
&self,
node_id: NodeId,
key_purpose: KeyPurpose,
version: RegistryVersion
) -> RegistryClientResult<PublicKeyProto>

fn get_threshold_signing_public_key_for_subnet(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<ThresholdSigPublicKey>

fn get_tls_certificate(
&self,
node_id: NodeId,
version: RegistryVersion
) -> RegistryClientResult<X509PublicKeyCert>

fn get_initial_dkg_transcripts(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientVersionedResult<DkgTranscripts>

Returns initial DKG key material for the subnet and the registry version, at which this key material was inserted. This registry version will be used in the genesis summary.

Loading content...