Trait ic_registry_client::helper::subnet::SubnetRegistry

pub trait SubnetRegistry {
    fn get_subnet_record(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<SubnetRecord>;

    fn get_root_subnet_id(
        &self, 
        version: RegistryVersion
    ) -> RegistryClientResult<SubnetId>;

    fn get_node_ids_on_subnet(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<Vec<NodeId>>;

    fn get_subnet_size(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<usize>;

    fn get_ingress_message_settings(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<IngressMessageSettings>;

    fn get_gossip_config(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<Option<GossipConfig>>;

    fn get_notarization_delay_settings(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<NotarizationDelaySettings>;

    fn get_dkg_dealings_per_block(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<usize>;

    fn get_dkg_interval_length(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<Height>;

    fn get_is_halted(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<bool>;

    fn get_replica_version(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<ReplicaVersion>;

    fn get_replica_version_record(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<ReplicaVersionRecord>;

    fn get_replica_version_record_from_version_id(
        &self, 
        replica_version_id: &ReplicaVersion, 
        version: RegistryVersion
    ) -> RegistryClientResult<ReplicaVersionRecord>;

    fn get_subnet_record_registry_version(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<RegistryVersion>;

    fn get_listed_subnet_for_node_id(
        &self, 
        node_id: NodeId, 
        version: RegistryVersion
    ) -> RegistryClientResult<(SubnetId, SubnetRecord)>;

    fn get_all_listed_subnet_records(
        &self, 
        version: RegistryVersion
    ) -> RegistryClientResult<Vec<(SubnetId, SubnetRecord)>>;

    fn get_cup_contents(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientVersionedResult<CatchUpPackageContents>;
}

A helper trait that wraps a RegistryClient and provides utility methods for querying subnet information.

Required methods

fn get_subnet_record(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<SubnetRecord>

fn get_root_subnet_id(
&self,
version: RegistryVersion
) -> RegistryClientResult<SubnetId>

fn get_node_ids_on_subnet(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<Vec<NodeId>>

fn get_subnet_size(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<usize>

fn get_ingress_message_settings(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<IngressMessageSettings>

Returns ingress message settings.

fn get_gossip_config(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<Option<GossipConfig>>

Returns gossip config

fn get_notarization_delay_settings(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<NotarizationDelaySettings>

Returns notarization delay settings:

the unit delay for blockmaker;

the initial delay for notary, to give time to rank-0 block propagation.

fn get_dkg_dealings_per_block(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<usize>

Returns the upper bound for the number of dealings we allow in a block.

fn get_dkg_interval_length(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<Height>

Returns the length of all DKG intervals for the given subnet. The interval length is the number of rounds, following the summary block, where dealers exchange their dealings.

fn get_is_halted(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<bool>

Returns whether the subnet record instructs the subnet to halt

fn get_replica_version(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<ReplicaVersion>

Return the ReplicaVersion as recorded in the subnet record at the given height.

fn get_replica_version_record(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<ReplicaVersionRecord>

Return the ReplicaVersionRecord as recorded in the subnet record at the given height.

fn get_replica_version_record_from_version_id(
&self,
replica_version_id: &ReplicaVersion,
version: RegistryVersion
) -> RegistryClientResult<ReplicaVersionRecord>

fn get_subnet_record_registry_version(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<RegistryVersion>

Return the RegistryVersion at which the SubnetRecord for the provided SubnetId was last updated.

fn get_listed_subnet_for_node_id(
&self,
node_id: NodeId,
version: RegistryVersion
) -> RegistryClientResult<(SubnetId, SubnetRecord)>

fn get_all_listed_subnet_records(
&self,
version: RegistryVersion
) -> RegistryClientResult<Vec<(SubnetId, SubnetRecord)>>

fn get_cup_contents(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientVersionedResult<CatchUpPackageContents>

Get the necessary material to construct a genesis/recovery CUP for the given subnet

Loading content...