Trait ic_registry_client::helper::subnet::SubnetTransportRegistry

pub trait SubnetTransportRegistry {
    fn get_subnet_transport_infos(
        &self, 
        subnet_id: SubnetId, 
        version: RegistryVersion
    ) -> RegistryClientResult<Vec<(NodeId, NodeRecord)>>;
}

Helper methods primarily used in transport/p2p where both, where transport information for an entire subnetwork are often needed.

Required methods

fn get_subnet_transport_infos(
&self,
subnet_id: SubnetId,
version: RegistryVersion
) -> RegistryClientResult<Vec<(NodeId, NodeRecord)>>

Return a list of pairs containing the node id and corresponding node record for each node on subnetwork with subnet_id.

As the transport information is stored individually for each node, this method performs n+1 requests, where n is the number of nodes on the network. Potential inconsistencies are resolved as follows:

Ok(None) is return if the a request for the subnet member list returns Ok(None), or if any of the requests for transport information return Ok(None).

Err(_) if the request for subnet membership fails.

The method panics in all other cases.

Panics

If the membership list for a subnet can be retrieved, but one of the requests for a node contained in the membership list fails, the method panics.

Loading content...