Module ic_registry_client::nns_registry_replicator

NNS Registry Replicator

(1) It polls one of the NNS Nodes for registry updates on a regular basis, verifies the response using the public key configured in the registry and applies the received changelog to the Registry Local Store.

(2) In case of a "switch-over" or starting a new independent NNS subnet, the NNS Registry Replicator modifies the Registry Local Store before rebooting:

Consider the registry of the «parent» IC instance as the source registry. Let subnet_record be a subnet record (in the source registry) with subnet_record.start_as_nns set to true. Let v be the registry version at which subnet_record was added to the registry (i.e. the smallest v for which subnet_record exists). Create a fresh (target) registry state that contains all versions up to and including v-1. Add version v, but with the following changes:

subnet_record.start_as_nns is unset on all subnet records

nns_subnet_id set to the new nns subnet id

subnet_list: contains only the nns_subnet_id

routing table: consists of a single entry that maps the same range of canister ids that was mapped to the NNS in the source registry to the subnet id obtained from subnet record

Concurrency

This is the only component that writes to the Registry Local Store. While individual changelog entries are stored atomically when replicating the registry, the switch-over is not atomic. This is the reason why the switch-over is handled in this component.

Structs

NnsRegistryReplicator